Rework MG poisson structures
The big advantage of Geometric MultiGrid is that the data has a strong spacial relationship - it isn't just a matrix and vector. To capture this, but still interface with LinearAlgebra packages, I created a few wrapper structs.

FieldVec (maybe rename GeomVec?) wraps a padded multidimensional array in a Vector. The type also defines a CartesianIndices array demarking the non-padded region. Linear indexing `(i::Int)` acts like any other vector as a fallback. CartesianIndexing accesses the underlying parent array for fast geometric referencing.

Poisson wraps the lower diagonal and main diagonal multidimensional arrays in a Matrix. Standard matrix indexing (i::Int, j::Int) used as a fallback, but the geometric parent arrays enable optimized `mul!` and `dot`.

This is all tested and working with `LinearAlgebra` and `IterativeSolvers`, which is great.