Add tuned! as default smoother
Played around with AD and was able to construct a simple parameterized local pseudo-inverse which is much faster and more effective than Gauss-Sidel. Total speed up is around x2 for Float32.

The coefficients have been tuned using guassian dipole on a 32x32 grid, but found to generalize well to other grid sizes and the uniform pressure gradient case.