Replace `tuned!` with `pseudo!`
Created type `PseudoInv <: FieldMatrix` to hold the tuned smoother matrix. The creation of the psuedo-inverse matrix coefficients adds overhead cost in time and memory. But once created, running `pseudo!` is much faster than `GS!`.

The default constructor for `PseudoInv` has been tuned using thousands of 2D examples of moving circles. It roughly doubles the speed of `mg!` for those cases, as well as the dipole and constant gradient cases.